requirements.txt
# Exported from the parnet_clean environment on 2026-04-09.
# Install ir_toolkit separately from GitHub as documented in README.md.
# Omitted machine-local dependency: rinalmo @
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
argcomplete==3.6.2
argh==0.31.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
array_record==0.8.1
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
biopython==1.84
bleach==6.2.0
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.8.2
captum==0.7.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
Cython==3.1.5
datasets==3.6.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
dm-tree==0.1.9
einops==0.8.1
et_xmlfile==2.0.0
etils==1.13.0
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.2.10
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.12.0
-e git+https://github.com/melonheader/gapitoo.git@99e7f04b48b9a5a5e8e5f5dcb08384363f4806ea#egg=gapitoo
gast==0.6.0
gdown==5.2.0
gffutils==0.13
gin-config==0.5.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_resources==6.5.2
intervaltree==3.1.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.11.3
kiwisolver==1.4.9
lark==1.2.2
lets-plot==4.7.2
libclang==18.1.1
lightning==2.5.4
lightning-fabric==2.5.1
lightning-utilities==0.14.3
llvmlite==0.42.0
logomaker==0.8.7
lxml==6.0.2
mappy==2.30
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
ml-dtypes==0.3.2
ml_collections==1.1.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
namex==0.1.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ncls==0.0.70
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
notebook==7.4.5
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.127
oauthlib==3.3.1
openpyxl==3.1.5
openspliceai==0.0.4
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==25.0
palettable==3.3.3
pandas==2.3.2
pandocfilters==1.5.1
parnet @ git+https://github.com/lambosaur/parnet.git@2e142786ffcec7ea5e23a28e1ec916c7524e6069
parso==0.8.5
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prometheus_client==0.22.1
promise==2.3
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.21.12
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybedtools==0.12.0
pyBigWig==0.3.24
pycparser==2.22
pyfaidx==0.9.0.3
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
pypng==0.20220715.0
pyranges==0.1.4
pysam==0.23.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
sequence-models==1.8.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sorted_nearest==0.0.41
sortedcontainers==2.4.0
soupsieve==2.8
stack-data==0.6.3
statsmodels==0.14.5
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-datasets==4.9.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.8.1
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.5
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
ViennaRNA==2.7.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
